requirements.txt
# Core dependencies
pandas
numpy>=1.24
tqdm>=4.60
pyyaml>=6.0
scikit-learn
requests

# KG and matching (optional but recommended)
faiss-cpu>=1.7
sentence-transformers>=2.0
rapidfuzz>=3.0
torch

# Lexicon loading
json5>=0.9

# LLM integration (install at least one)
google-genai>=0.3
# openai>=1.0

# Validation
pydantic>=2.0
jsonschema>=4.0

# Testing
pytest>=7.0
